remove components folder and organize settings
