+++ app/features/settings/accounts/create-account.tsx
@@ -1,6 +1,6 @@
 import { PageContent } from '~/components/page';
-import { SettingsNavButton } from '~/features/settings/components/settings-nav-button';
-import { SettingsViewHeader } from '~/features/settings/components/settings-view-header';
+import { SettingsNavButton } from '~/features/settings/ui/settings-nav-button';
+import { SettingsViewHeader } from '~/features/settings/ui/settings-view-header';
 
 export default function CreateAccount() {
   return (
