+++ app/features/settings/accounts/nwc.tsx
@@ -1,4 +1,4 @@
-import { AddLightningWalletForm } from '~/features/settings/components/add-ln-wallet-form';
+import { AddLightningWalletForm } from '~/features/settings/accounts/add-ln-wallet-form';
 
 export default function AddNWCAccountView() {
   return <AddLightningWalletForm />;
