+++ app/features/settings/appearance.tsx
@@ -1,5 +1,5 @@
 import { PageContent } from '~/components/page';
-import { SettingsViewHeader } from '~/features/settings/components/settings-view-header';
+import { SettingsViewHeader } from '~/features/settings/ui/settings-view-header';
 import { ColorModeToggle } from '~/features/theme/color-mode-toggle';
 import { useTheme } from '~/features/theme/use-theme';
