+++ app/features/settings/profile/share-profile-qr.tsx
