+++ app/routes/_protected.settings.accounts._index.tsx
@@ -1,5 +1,5 @@
 import { Page } from '~/components/page';
-import AllAccounts from '~/features/settings/all-accounts';
+import AllAccounts from '~/features/settings/accounts/all-accounts';
 
 export default function AllAccountsPage() {
   return (
