+++ app/routes/_protected.settings.accounts.create.tsx
@@ -1,5 +1,5 @@
 import { Page } from '~/components/page';
-import CreateAccount from '~/features/settings/create-account';
+import CreateAccount from '~/features/settings/accounts/create-account';
 
 export default function CreateAccountPage() {
   return (
