+++ app/routes/_protected.settings.appearance.tsx
@@ -1,5 +1,5 @@
 import { Page } from '~/components/page';
-import AppearanceSettings from '~/features/settings/appearance-settings';
+import AppearanceSettings from '~/features/settings/appearance';
 
 export default function AppearanceSettingsPage() {
   return (
