+++ app/routes/_protected.settings.profile.edit.tsx
@@ -1,5 +1,5 @@
 import { Page } from '~/components/page';
-import EditProfile from '~/features/settings/edit-profile';
+import EditProfile from '~/features/settings/profile/edit-profile';
 
 export default function EditProfilePage() {
   return (
