+++ app/routes/_protected.settings.qr.tsx
@@ -1,5 +1,5 @@
 import { Page } from '~/components/page';
-import ProfileQR from '~/features/settings/profile-qr';
+import ProfileQR from '~/features/settings/profile/share-profile-qr';
 
 export default function ProfileQRPage() {
   return (
